+++ plugins/guests/coreos/plugin.rb
@@ -20,6 +20,11 @@ module VagrantPlugins
         require_relative "cap/configure_networks"
         Cap::ConfigureNetworks
       end
+
+      guest_capability("coreos", "docker_daemon_running") do
+        require_relative "cap/docker"
+        Cap::Docker
+      end
     end
   end
 end